+++ exchanges_wrapper/ftx_parser.py
@@ -80,6 +80,19 @@ class OrderBook:
         return self
 
 
+def on_balance_update(res: {}) -> {}:
+    ftx_time = res.get('time')
+    _time = int(datetime.datetime.strptime(ftx_time.split('+')[0], TIMESTAMP_PATTERN).timestamp() * 1000)
+    balance = {
+        'e': 'balanceUpdate',
+        'E': _time,
+        'a': res.get('coin'),
+        'd': str(res.get('size')),
+        'T': int(time.time() * 1000)
+    }
+    return balance
+
+
 def on_funds_update(res: []) -> {}:
     binance_funds = {
         'e': 'outboundAccountPosition',
@@ -147,6 +160,12 @@ def exchange_info(res: []) -> {}:
                 "applyToMarket": True,
                 "avgPriceMins": 5
             }
+            _percent_price = {
+                "filterType": "PERCENT_PRICE",
+                "multiplierUp": "5",
+                "multiplierDown": "0.2",
+                "avgPriceMins": 5
+            }
 
             symbol = {
                 "symbol": _symbol,
@@ -166,7 +185,7 @@ def exchange_info(res: []) -> {}:
                 "cancelReplaceAllowed": False,
                 "isSpotTradingAllowed": True,
                 "isMarginTradingAllowed": False,
-                "filters": [_price_filter, _lot_size, _min_notional],
+                "filters": [_price_filter, _lot_size, _min_notional, _percent_price],
                 "permissions": ["SPOT"],
             }
             symbols.append(symbol)