+++ exchanges_wrapper/okx_parser.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
 """
-Parser for convert Huobi REST API/WSS response to Binance like result
+Parser for convert OKX REST API/WSS response to Binance like result
 """
 import time
 from decimal import Decimal
@@ -10,83 +10,69 @@ import logging
 logger = logging.getLogger('exch_srv_logger')
 
 
-def on_balance_update(res: {}) -> {}:
-    balance = {
-        'e': 'balanceUpdate',
-        'E': res.get('transactTime'),
-        'a': res.get('currency').upper(),
-        'd': str(res.get('transactAmt')),
-        'T': int(time.time() * 1000)
-    }
-    return balance
+def fetch_server_time(res: []) -> {}:
+    if res:
+        return {'serverTime': int(res[0].get('ts'))}
 
 
-def fetch_server_time(res: {}) -> {}:
-    return {'serverTime': res}
-
-
-def exchange_info(server_time: int, trading_symbol: []) -> {}:
+def exchange_info(server_time: int, trading_symbol: [], tickers: []) -> {}:
     symbols = []
+    symbols_price = {}
+    for pair in tickers:
+        symbols_price[pair.get('instId').replace('-', '')] = Decimal(pair.get('last'))
+
     for market in trading_symbol:
-        if not market.get('underlying'):
-            _symbol = str(market.get("symbol")).upper()
-            _base_asset = str(market.get("base-currency")).upper()
-            _quote_asset = str(market.get("quote-currency")).upper()
-            _base_asset_precision = market.get('amount-precision')
-            # Filters var
-            _tick_size = 10**(-market.get('price-precision'))
-            _min_qty = market.get('min-order-amt')
-            _max_qty = market.get('max-order-amt')
-            _step_size = 10**(-market.get('amount-precision'))
-            _min_notional = market.get('min-order-value')
+        _symbol = market.get("instId").replace('-', '')
+        _base_asset = market.get("baseCcy")
+        _quote_asset = market.get("quoteCcy")
+        _base_asset_precision = len(market.get('lotSz')) - 2
+        # Filters var
+        _tick_size = market.get('tickSz')
+        _min_qty = market.get('minSz')
+        _max_qty = market.get('maxLmtSz')
+        _step_size = market.get('lotSz')
+        _min_notional = str(Decimal(_min_qty) * symbols_price.get(_symbol))
+        _price_filter = {
+            "filterType": "PRICE_FILTER",
+            "minPrice": str(_tick_size),
+            "maxPrice": "100000.00000000",
+            "tickSize": str(_tick_size)
+        }
+        _lot_size = {
+            "filterType": "LOT_SIZE",
+            "minQty": str(_min_qty),
+            "maxQty": str(_max_qty),
+            "stepSize": str(_step_size)
+        }
+        _min_notional = {
+            "filterType": "MIN_NOTIONAL",
+            "minNotional": str(_min_notional),
+            "applyToMarket": True,
+            "avgPriceMins": 0
+        }
 
-            _price_filter = {
-                "filterType": "PRICE_FILTER",
-                "minPrice": str(_tick_size),
-                "maxPrice": "100000.00000000",
-                "tickSize": str(_tick_size)
-            }
-            _lot_size = {
-                "filterType": "LOT_SIZE",
-                "minQty": str(_min_qty),
-                "maxQty": str(_max_qty),
-                "stepSize": str(_step_size)
-            }
-            _min_notional = {
-                "filterType": "MIN_NOTIONAL",
-                "minNotional": str(_min_notional),
-                "applyToMarket": True,
-                "avgPriceMins": 5
-            }
-            _percent_price = {
-                "filterType": "PERCENT_PRICE",
-                "multiplierUp": "5",
-                "multiplierDown": "0.2",
-                "avgPriceMins": 5
-            }
-
-            symbol = {
-                "symbol": _symbol,
-                "status": "TRADING",
-                "baseAsset": _base_asset,
-                "baseAssetPrecision": _base_asset_precision,
-                "quoteAsset": _quote_asset,
-                "quotePrecision": _base_asset_precision,
-                "quoteAssetPrecision": _base_asset_precision,
-                "baseCommissionPrecision": 8,
-                "quoteCommissionPrecision": 8,
-                "orderTypes": ["LIMIT", "MARKET"],
-                "icebergAllowed": False,
-                "ocoAllowed": False,
-                "quoteOrderQtyMarketAllowed": False,
-                "allowTrailingStop": False,
-                "cancelReplaceAllowed": False,
-                "isSpotTradingAllowed": True,
-                "isMarginTradingAllowed": False,
-                "filters": [_price_filter, _lot_size, _min_notional, _percent_price],
-                "permissions": ["SPOT"],
-            }
-            symbols.append(symbol)
+        symbol = {
+            "symbol": _symbol,
+            "status": "TRADING",
+            "baseAsset": _base_asset,
+            "baseAssetPrecision": _base_asset_precision,
+            "quoteAsset": _quote_asset,
+            "quotePrecision": _base_asset_precision,
+            "quoteAssetPrecision": _base_asset_precision,
+            "baseCommissionPrecision": 8,
+            "quoteCommissionPrecision": 8,
+            "orderTypes": ["LIMIT", "MARKET"],
+            "icebergAllowed": False,
+            "ocoAllowed": False,
+            "quoteOrderQtyMarketAllowed": False,
+            "allowTrailingStop": False,
+            "cancelReplaceAllowed": False,
+            "isSpotTradingAllowed": True,
+            "isMarginTradingAllowed": False,
+            "filters": [_price_filter, _lot_size, _min_notional],
+            "permissions": ["SPOT"],
+        }
+        symbols.append(symbol)
 
     _binance_res = {
         "timezone": "UTC",
@@ -107,19 +93,20 @@ def orders(res: [], response_type=None) -> []:
 
 
 def order(res: {}, response_type=None) -> {}:
-    symbol = res.get('symbol').upper()
-    order_id = res.get('id')
+    symbol = res.get('instId').replace('-', '')
+    order_id = int(res.get('ordId'))
     order_list_id = -1
-    client_order_id = res.get('client-order-id')
-    price = res.get('price', "0")
-    orig_qty = res.get('amount', "0")
-    executed_qty = res.get('filled-amount', res.get('field-amount', "0"))
-    cummulative_quote_qty = res.get('filled-cash-amount', res.get('field-cash-amount', "0"))
+    client_order_id = res.get('clOrdId')
+    price = res.get('px', "0")
+    orig_qty = res.get('sz', "0")
+    executed_qty = res.get('accFillSz')
+    avg_filled_price = res.get('avgPx') or "0"
+    cummulative_quote_qty = str(Decimal(executed_qty) * Decimal(avg_filled_price))
     orig_quote_order_qty = str(Decimal(orig_qty) * Decimal(price))
     #
-    if res.get('state') in ('canceled', 'partial-canceled'):
+    if res.get('state') == 'canceled':
         status = 'CANCELED'
-    elif res.get('state') == 'partial-filled':
+    elif res.get('state') == 'partially_filled':
         status = 'PARTIALLY_FILLED'
     elif res.get('state') == 'filled':
         status = 'FILLED'
@@ -128,11 +115,11 @@ def order(res: {}, response_type=None) -> {}:
     #
     _type = "LIMIT"
     time_in_force = "GTC"
-    side = 'BUY' if 'buy' in res.get('type') else 'SELL'
+    side = 'BUY' if 'buy' in res.get('side') else 'SELL'
     stop_price = '0.0'
     iceberg_qty = '0.0'
-    _time = res.get('created-at')
-    update_time = res.get('canceled-at') or res.get('finished-at') or _time
+    _time = int(res.get('cTime'))
+    update_time = int(res.get('uTime'))
     is_working = True
     #
     if response_type:
@@ -192,25 +179,13 @@ def order(res: {}, response_type=None) -> {}:
     return binance_order
 
 
-def account_information(res: {}) -> {}:
+def account_information(res: [], u_time: str) -> {}:
     balances = []
-    res[:] = [i for i in res if i.get('balance') != '0']
-    assets = {}
-    for balance in res:
-        asset = balance.get('currency')
-        asset_i = assets.get(asset, {})
-        if balance.get('available'):
-            asset_i.setdefault('available', balance.get('available'))
-        else:
-            asset_i.setdefault('frozen', balance.get('balance', '0'))
-        assets.update({asset: asset_i})
-    for asset in assets:
-        free = assets.get(asset, {}).get('available', '0')
-        locked = assets.get(asset, {}).get('frozen', '0')
+    for asset in res:
         _binance_res = {
-            "asset": asset.upper(),
-            "free": free,
-            "locked": locked,
+            "asset": asset.get('ccy'),
+            "free": asset.get('availBal'),
+            "locked": asset.get('frozenBal'),
         }
         balances.append(_binance_res)
 
@@ -222,7 +197,8 @@ def account_information(res: {}) -> {}:
       "canTrade": True,
       "canWithdraw": False,
       "canDeposit": False,
-      "updateTime": int(time.time() * 1000),
+      "brokered": False,
+      "updateTime": int(u_time),
       "accountType": "SPOT",
       "balances": balances,
       "permissions": [
@@ -237,6 +213,9 @@ def order_book(res: {}) -> {}:
     binance_order_book.setdefault('bids', res.get('bids'))
     binance_order_book.setdefault('asks', res.get('asks'))
     return binance_order_book
+
+
+###############################################################################
 
 
 def order_book_ws(res: {}, symbol: str) -> {}:
@@ -495,3 +474,108 @@ def account_trade_list(res: []) -> []:
         }
         binance_trade_list.append(binance_trade)
     return binance_trade_list
+###############################################################################
+
+
+def get_symbols(symbols_details: []) -> str:
+    symbols = []
+    res = ",t"
+    for symbol_details in symbols_details:
+        symbol = symbol_details['pair']
+        if 'f0' not in symbol:
+            symbols.append(symbol.upper())
+    return f"t{res.join(symbols)}"
+
+
+def tick_size(precision, _price):
+    x = int(_price)
+    _price = str(_price)
+    if '.' not in _price:
+        _price += ".0"
+    k = len(_price.split('.')[1])
+    x = len(_price.split('.')[0]) if k and x else 0
+    if k + x - precision > 0:
+        k = precision - x
+    elif k + x - precision < 0:
+        k += precision - x - k
+    res = (1 / 10 ** k) if k else 1
+    return res
+
+
+def symbol_name(_pair: str) -> ():
+    if ':' in _pair:
+        pair = _pair.replace(':', '').upper()
+        base_asset = _pair.split(':')[0].upper()
+        quote_asset = _pair.split(':')[1].upper()
+    else:
+        pair = _pair.upper()
+        base_asset = _pair[0:3].upper()
+        quote_asset = _pair[3:].upper()
+    return pair, base_asset, quote_asset
+
+
+def on_order_trade(res: [], executed_qty: str) -> {}:
+    # print(f"on_order_trade.res: {res}")
+    side = 'BUY' if res[4] > 0 else 'SELL'
+    #
+    status = 'PARTIALLY_FILLED'
+    #
+    last_executed_quantity = str(abs(res[4]))
+    last_executed_price = str(res[5])
+    last_quote_asset = str(Decimal(last_executed_quantity) * Decimal(last_executed_price))
+    msg_binance = {
+        "e": "executionReport",
+        "E": res[2],
+        "s": res[1][1:].replace(':', ''),
+        "c": str(res[11]),
+        "S": side,
+        "o": "LIMIT",
+        "f": "GTC",
+        "q": "0.0",
+        "p": str(res[7]),
+        "P": "0.00000000",
+        "F": "0.00000000",
+        "g": -1,
+        "C": "NEW",
+        "x": "TRADE",
+        "X": status,
+        "r": "NONE",
+        "i": res[3],
+        "l": last_executed_quantity,
+        "z": executed_qty,
+        "L": last_executed_price,
+        "n": str(res[9]),
+        "N": res[10],
+        "T": res[2],
+        "t": res[0],
+        "I": 123456789,
+        "w": True,
+        "m": bool(res[8] == 1),
+        "M": False,
+        "O": res[2],
+        "Z": "0.0",
+        "Y": last_quote_asset,
+        "Q": "0.0"
+    }
+    return msg_binance
+
+
+def funding_wallet(res: []) -> []:
+    balances = []
+    for balance in res:
+        if balance[0] in ('exchange', 'funding'):
+            total = str(balance[2] or 0.0)
+            free = str(balance[4] or 0.0)
+            locked = str(Decimal(total) - Decimal(free))
+            if float(total):
+                _binance_res = {
+                    "asset": balance[1],
+                    "free": free,
+                    "locked": locked,
+                    "freeze": "0",
+                    "withdrawing": "0",
+                    "btcValuation": "0.0",
+                }
+                balances.append(_binance_res)
+
+    return balances